Add new branch feature. Structure components into directories. Change paths of components.
