+++ frontend/src/components/Dashboards/Company/CompanySort.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useForm } from "react-hook-form";
-import { getCompanyJobs } from "../slices/companyJobsSlice";
+import { getCompanyJobs } from "../../../slices/companyJobsSlice";
 import { useDispatch, useSelector } from "react-redux";
 import { FaSortAlphaDown } from "react-icons/fa";
