+++ frontend/src/components/Dashboards/Company/CreateJobs.jsx
@@ -1,11 +1,11 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import { set, useForm } from "react-hook-form";
+import { useForm } from "react-hook-form";
 import axios from "axios";
 import { useAuth0 } from "@auth0/auth0-react";
 import { useDispatch } from "react-redux";
-import { getCompanyJobs } from "../slices/companyJobsSlice";
-import CharactersUsed from "./CharactersUsed";
+import { getCompanyJobs } from "../../../slices/companyJobsSlice";
+import CharactersUsed from "../../Global/CharactersUsed";
 
 const Style = styled.section`
   position: absolute;
