+++ frontend/src/components/Dashboards/Company/WriteMessage.jsx
@@ -1,22 +1,20 @@
 import React from "react";
 import styled from "styled-components";
 import { useForm } from "react-hook-form";
-import axios from "axios";
 import { useAuth0 } from "@auth0/auth0-react";
-import { useParams } from "react-router-dom";
+import axios from "axios";
+import { getMessages } from "../../../slices/messagesSlice";
 import { useDispatch } from "react-redux";
-import { getApps } from "../slices/appsSlice";
-import { useNavigate } from "react-router-dom";
-import CharactersUsed from "./CharactersUsed";
+import CharactersUsed from "../../Global/CharactersUsed";
 
 const Style = styled.section`
   position: fixed;
-  background-color: rgba(0, 0, 0, 0.75);
-  z-index: 10;
   top: 0;
-  width: 100%;
-  min-height: 100vh;
   padding-top: 5rem;
+  background-color: rgba(0, 0, 0, 0.75);
+  z-index: 100;
+  width: 100%;
+  min-height: 100dvh;
 
   input,
   textarea {
@@ -32,16 +30,12 @@ const Style = styled.section`
     border: 1px solid #ccc;
   }
 
-  ul {
-    margin-left: 2rem;
-  }
-
   @media screen and (max-width: 767px) {
     padding-top: 2rem;
   }
 `;
 
-const ApplicationForm = ({ setIsApplication }) => {
+const WriteMessage = ({ setIsMessageOpen, companyApps, selectedApp }) => {
   const {
     register,
     handleSubmit,
@@ -50,24 +44,23 @@ const ApplicationForm = ({ setIsApplication }) => {
   } = useForm();
 
   const { getAccessTokenSilently } = useAuth0();
-  const param = useParams();
   const dispatch = useDispatch();
-  const navigate = useNavigate();
   const baseUrl = import.meta.env.VITE_BASE_URL;
 
   const onSubmit = async (data) => {
     console.log(data);
+    console.log(companyApps);
     try {
-      data.job_id = param.jobId;
+      data.app_id = companyApps[selectedApp].app?.app_id;
       const token = await getAccessTokenSilently();
-      const response = await axios.post(`${baseUrl}/apply`, data, {
+      const response = await axios.post(`${baseUrl}/createMessage`, data, {
         headers: {
           Authorization: `Bearer ${token}`,
         },
       });
       console.log(response.data);
-      dispatch(getApps({ payload: response.data }));
-      navigate(`/dashboard`);
+      dispatch(getMessages({ messages: response.data, isLoading: false }));
+      setIsMessageOpen(false);
     } catch (error) {
       console.error("Error calling API:", error);
       console.log(error.response.data);
@@ -78,20 +71,20 @@ const ApplicationForm = ({ setIsApplication }) => {
     <Style>
       <button
         className="btn block m-auto border-0 text-white capitalize my-4 bg-red-500 hover:bg-red-200"
-        onClick={() => setIsApplication(false)}
+        onClick={() => setIsMessageOpen(false)}
+        aria-label="close"
       >
         CLOSE
       </button>
-
       <form
         onSubmit={handleSubmit(onSubmit)}
         action=""
         className="flex flex-col bg-neutral m-auto p-10 max-md:w-[327px] 
         max-md:p-6 md:w-[690px] xl:w-[1100px] shadow rounded-xl"
       >
-        <h2 className="text-center font-bold">Write Application</h2>
+        <h2 className="text-center font-bold">Write Message</h2>
         <div className="flex flex-col gap-4 justify-center items-center my-4">
-          <label htmlFor="content">Application Text</label>
+          <label htmlFor="content">Message Text</label>
           <textarea
             className="bg-neutral text-primary"
             type="text"
@@ -101,8 +94,13 @@ const ApplicationForm = ({ setIsApplication }) => {
               minLength: 50,
               maxLength: 500,
             })}
-            placeholder="Write your application here (min 50 and max 500 characters)"
+            placeholder="Write your message here (min 50 and max 500 characters)"
             aria-invalid={errors.content ? "true" : "false"}
+          />
+          <CharactersUsed
+            charactersUsed={watch("content")?.length || 0}
+            maxCharacters={500}
+            offset={{ small: 0, medium: 0, large: 0 }}
           />
           {errors.content?.type === "required" && (
             <p className="text-red-500" role="alert">
@@ -120,14 +118,9 @@ const ApplicationForm = ({ setIsApplication }) => {
             </p>
           )}
         </div>
-        <CharactersUsed
-          charactersUsed={watch("content")?.length || 0}
-          maxCharacters={500}
-          offset={{ small: 0, medium: 0, large: 0 }}
-        />
         <input
           type="submit"
-          value="submit application"
+          value="submit message"
           className="btn my-4 duration-0 capitalize text-white bg-accent max-md:max-w-full 
           max-w-[10rem] self-center hover:bg-info border-none"
         />
@@ -136,4 +129,4 @@ const ApplicationForm = ({ setIsApplication }) => {
   );
 };
 
-export default ApplicationForm;
+export default WriteMessage;
