+++ frontend/src/components/Dashboards/Private/ReplyMessage.jsx
@@ -3,16 +3,14 @@ import styled from "styled-components";
 import { useForm } from "react-hook-form";
 import { useAuth0 } from "@auth0/auth0-react";
 import axios from "axios";
-import { getMessages } from "../slices/messagesSlice";
-import { useDispatch } from "react-redux";
-import CharactersUsed from "./CharactersUsed";
+import CharactersUsed from "../../Global/CharactersUsed";
 
 const Style = styled.section`
   position: fixed;
   top: 0;
   padding-top: 5rem;
   background-color: rgba(0, 0, 0, 0.75);
-  z-index: 100;
+  z-index: 1000;
   width: 100%;
   min-height: 100dvh;
 
@@ -35,7 +33,7 @@ const Style = styled.section`
   }
 `;
 
-const WriteMessage = ({ setIsMessageOpen, companyApps, selectedApp }) => {
+const ReplyMessage = ({ setIsReplyOpen, selectedMessage }) => {
   const {
     register,
     handleSubmit,
@@ -44,23 +42,21 @@ const WriteMessage = ({ setIsMessageOpen, companyApps, selectedApp }) => {
   } = useForm();
 
   const { getAccessTokenSilently } = useAuth0();
-  const dispatch = useDispatch();
   const baseUrl = import.meta.env.VITE_BASE_URL;
 
   const onSubmit = async (data) => {
     console.log(data);
-    console.log(companyApps);
     try {
-      data.app_id = companyApps[selectedApp].app?.app_id;
+      console.log(selectedMessage);
+      data.message_id = selectedMessage;
       const token = await getAccessTokenSilently();
-      const response = await axios.post(`${baseUrl}/createMessage`, data, {
+      const response = await axios.post(`${baseUrl}/createReply`, data, {
         headers: {
           Authorization: `Bearer ${token}`,
         },
       });
       console.log(response.data);
-      dispatch(getMessages({ messages: response.data, isLoading: false }));
-      setIsMessageOpen(false);
+      setIsReplyOpen(false);
     } catch (error) {
       console.error("Error calling API:", error);
       console.log(error.response.data);
@@ -71,7 +67,7 @@ const WriteMessage = ({ setIsMessageOpen, companyApps, selectedApp }) => {
     <Style>
       <button
         className="btn block m-auto border-0 text-white capitalize my-4 bg-red-500 hover:bg-red-200"
-        onClick={() => setIsMessageOpen(false)}
+        onClick={() => setIsReplyOpen(false)}
         aria-label="close"
       >
         CLOSE
@@ -82,9 +78,9 @@ const WriteMessage = ({ setIsMessageOpen, companyApps, selectedApp }) => {
         className="flex flex-col bg-neutral m-auto p-10 max-md:w-[327px] 
         max-md:p-6 md:w-[690px] xl:w-[1100px] shadow rounded-xl"
       >
-        <h2 className="text-center font-bold">Write Message</h2>
+        <h2 className="text-center font-bold">Write Reply</h2>
         <div className="flex flex-col gap-4 justify-center items-center my-4">
-          <label htmlFor="content">Message Text</label>
+          <label htmlFor="content">Reply Text</label>
           <textarea
             className="bg-neutral text-primary"
             type="text"
@@ -120,6 +116,7 @@ const WriteMessage = ({ setIsMessageOpen, companyApps, selectedApp }) => {
         </div>
         <input
           type="submit"
+          aria-label="submit message"
           value="submit message"
           className="btn my-4 duration-0 capitalize text-white bg-accent max-md:max-w-full 
           max-w-[10rem] self-center hover:bg-info border-none"
@@ -129,4 +126,4 @@ const WriteMessage = ({ setIsMessageOpen, companyApps, selectedApp }) => {
   );
 };
 
-export default WriteMessage;
+export default ReplyMessage;
