+++ frontend/src/components/Dashboards/Private/ViewAppDetails.jsx
