+++ frontend/src/components/Dashboards/Profile/EditProfile.jsx
@@ -4,7 +4,7 @@ import { useForm } from "react-hook-form";
 import { useAuth0 } from "@auth0/auth0-react";
 import axios from "axios";
 import { useDispatch } from "react-redux";
-import { setUser } from "../slices/userSlice";
+import { setUser } from "../../../slices/userSlice";
 
 const Style = styled.section`
   position: fixed;
