+++ frontend/src/components/Global/CharactersUsed.jsx
