+++ frontend/src/components/Global/Footer.jsx
