+++ frontend/src/components/Inner/JobFooter.jsx
